% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3podemo-package.R
\docType{package}
\name{d3podemo}
\alias{d3podemo-package}
\alias{d3podemo}
\title{D3po + Tabler Demo}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
