% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_area2.R
\name{mod_area2_plot}
\alias{mod_area2_plot}
\title{Area2 (stacked) plot module}
\usage{
mod_area2_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Area2 (stacked) plot module
}
