% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_area4.R
\name{mod_area4_plot}
\alias{mod_area4_plot}
\title{Area chart: stacked example}
\usage{
mod_area4_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Area chart: stacked example
}
