% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_area1.R
\name{mod_area_plot}
\alias{mod_area_plot}
\title{Area chart: non-stacked}
\usage{
mod_area_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Area chart: non-stacked
}
