% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar1.R
\name{mod_bar1_plot}
\alias{mod_bar1_plot}
\title{Bar chart: Vertical Bars}
\usage{
mod_bar1_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame for bar aggregation (expects pre-aggregated if passed)}
}
\value{
d3po widget
}
\description{
Bar chart: Vertical Bars
}
