% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar2.R
\name{mod_bar2_plot}
\alias{mod_bar2_plot}
\title{Bar chart: Horizontal Bars}
\usage{
mod_bar2_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Bar chart: Horizontal Bars
}
