% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar3.R
\name{mod_bar3_plot}
\alias{mod_bar3_plot}
\title{Bar3 plot module (stacked vertical)}
\usage{
mod_bar3_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Bar3 plot module (stacked vertical)
}
