% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar4.R
\name{mod_bar4_plot}
\alias{mod_bar4_plot}
\title{Bar chart: Stacked Horizontal Bars}
\usage{
mod_bar4_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Bar chart: Stacked Horizontal Bars
}
