% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar_custom1.R
\name{mod_bar_custom1_plot}
\alias{mod_bar_custom1_plot}
\title{Bar custom1 plot module}
\usage{
mod_bar_custom1_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Bar custom1 plot module
}
