% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar_custom2.R
\name{mod_bar_custom2_plot}
\alias{mod_bar_custom2_plot}
\title{Bar custom2 plot module}
\usage{
mod_bar_custom2_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Bar custom2 plot module
}
