% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar_custom.R
\name{mod_bar_custom_plot}
\alias{mod_bar_custom_plot}
\title{Custom Bar plot: Mean weight by type/color}
\usage{
mod_bar_custom_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Custom Bar plot: Mean weight by type/color
}
