% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box3.R
\name{mod_box3_plot}
\alias{mod_box3_plot}
\title{Box3 plot module (log weight)}
\usage{
mod_box3_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Box3 plot module (log weight)
}
