% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box6.R
\name{mod_box6_plot}
\alias{mod_box6_plot}
\title{Box plot: Weight Distribution by Type}
\usage{
mod_box6_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Box plot: Weight Distribution by Type
}
