% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box_custom1.R
\name{mod_box_custom1_plot}
\alias{mod_box_custom1_plot}
\title{Box custom1 plot module}
\usage{
mod_box_custom1_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Box custom1 plot module
}
