% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box_custom2.R
\name{mod_box_custom2_plot}
\alias{mod_box_custom2_plot}
\title{Custom Box plot: custom background, axis and tooltip colors}
\usage{
mod_box_custom2_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Custom Box plot: custom background, axis and tooltip colors
}
