% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box_custom.R
\name{mod_box_custom_plot}
\alias{mod_box_custom_plot}
\title{Custom Box plot: Weight in grams (custom example)}
\usage{
mod_box_custom_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Custom Box plot: Weight in grams (custom example)
}
