% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geomap1.R
\name{mod_geomap1_plot}
\alias{mod_geomap1_plot}
\title{Geomap: South America}
\usage{
mod_geomap1_plot(data = NULL)
}
\arguments{
\item{data}{data.frame (optional)}
}
\value{
d3po widget
}
\description{
Geomap: South America
}
