% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geomap2.R
\name{mod_geomap2_plot}
\alias{mod_geomap2_plot}
\title{Geomap2 plot module}
\usage{
mod_geomap2_plot(data)
}
\arguments{
\item{data}{data.frame of region ids}
}
\value{
d3po widget
}
\description{
Geomap2 plot module
}
