% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line3.R
\name{mod_line3_plot}
\alias{mod_line3_plot}
\title{Line chart example}
\usage{
mod_line3_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Line chart example
}
