% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line_area_custom.R
\name{mod_line_area_custom_plot}
\alias{mod_line_area_custom_plot}
\title{Line/Area custom plot module}
\usage{
mod_line_area_custom_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Line/Area custom plot module
}
