% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_custom.R
\name{mod_network_custom_plot}
\alias{mod_network_custom_plot}
\title{Network: custom tooltip example}
\usage{
mod_network_custom_plot(data = d3po::pokemon_network)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Network: custom tooltip example
}
