% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_fr.R
\name{mod_network_fr_plot}
\alias{mod_network_fr_plot}
\title{Network: FR layout}
\usage{
mod_network_fr_plot(data = d3po::pokemon_network)
}
\arguments{
\item{data}{data.frame (pokemon_network expected)}
}
\value{
d3po widget
}
\description{
Network: FR layout
}
