% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_kk_nocolors.R
\name{mod_network_kk_nocolors_plot}
\alias{mod_network_kk_nocolors_plot}
\title{Network: KK layout}
\usage{
mod_network_kk_nocolors_plot(data = d3po::pokemon_network)
}
\arguments{
\item{data}{data.frame (pokemon_network expected)}
}
\value{
d3po widget
}
\description{
Network: KK layout
}
