% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_kk.R
\name{mod_network_kk_plot}
\alias{mod_network_kk_plot}
\title{Network KK layout plot module}
\usage{
mod_network_kk_plot(data)
}
\arguments{
\item{data}{graph or nodes/edges}
}
\value{
d3po widget
}
\description{
Network KK layout plot module
}
