% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_manual.R
\name{mod_network_manual_plot}
\alias{mod_network_manual_plot}
\title{Network manual layout plot module}
\usage{
mod_network_manual_plot(data)
}
\arguments{
\item{data}{graph with layout}
}
\value{
d3po widget
}
\description{
Network manual layout plot module
}
