% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pie6.R
\name{mod_pie6_plot}
\alias{mod_pie6_plot}
\title{Pie chart: Full Pie}
\usage{
mod_pie6_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Pie chart: Full Pie
}
