% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pie_custom.R
\name{mod_pie_custom_plot}
\alias{mod_pie_custom_plot}
\title{Pie with custom labels and tooltip}
\usage{
mod_pie_custom_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Pie with custom labels and tooltip
}
