% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter1.R
\name{mod_scatter1_plot}
\alias{mod_scatter1_plot}
\title{Scatter: Height vs Weight}
\usage{
mod_scatter1_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Scatter: Height vs Weight
}
