% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter2.R
\name{mod_scatter2_plot}
\alias{mod_scatter2_plot}
\title{Scatter2 plot module}
\usage{
mod_scatter2_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Scatter2 plot module
}
