% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter3.R
\name{mod_scatter3_plot}
\alias{mod_scatter3_plot}
\title{Scatter3 plot module}
\usage{
mod_scatter3_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Scatter3 plot module
}
