% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter4.R
\name{mod_scatter4_plot}
\alias{mod_scatter4_plot}
\title{Scatter4 plot module}
\usage{
mod_scatter4_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Scatter4 plot module
}
