% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_custom.R
\name{mod_scatter_custom_plot}
\alias{mod_scatter_custom_plot}
\title{Scatter custom plot module}
\usage{
mod_scatter_custom_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Scatter custom plot module
}
