% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap1.R
\name{mod_treemap1_plot}
\alias{mod_treemap1_plot}
\title{Treemap1 plot module}
\usage{
mod_treemap1_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon or summary}
}
\value{
d3po widget
}
\description{
Treemap1 plot module
}
