% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap2.R
\name{mod_treemap2_plot}
\alias{mod_treemap2_plot}
\title{Treemap2 (two-level) plot module}
\usage{
mod_treemap2_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Treemap2 (two-level) plot module
}
