% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap3.R
\name{mod_treemap3_plot}
\alias{mod_treemap3_plot}
\title{Treemap3 plot module}
\usage{
mod_treemap3_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap3 plot module
}
