% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap4.R
\name{mod_treemap4_plot}
\alias{mod_treemap4_plot}
\title{Treemap4 plot module}
\usage{
mod_treemap4_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap4 plot module
}
