% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_custom2.R
\name{mod_treemap_custom2_plot}
\alias{mod_treemap_custom2_plot}
\title{Treemap custom2 module}
\usage{
mod_treemap_custom2_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap custom2 module
}
