% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_custom.R
\name{mod_treemap_custom_plot}
\alias{mod_treemap_custom_plot}
\title{Treemap custom module}
\usage{
mod_treemap_custom_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap custom module
}
