% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_onelevel_custom2.R
\name{mod_treemap_onelevel_custom2_plot}
\alias{mod_treemap_onelevel_custom2_plot}
\title{Treemap with custom labels (Rmd-style labels JS example)}
\usage{
mod_treemap_onelevel_custom2_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap with custom labels (Rmd-style labels JS example)
}
