% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_onelevel_custom.R
\name{mod_treemap_onelevel_custom_plot}
\alias{mod_treemap_onelevel_custom_plot}
\title{Treemap with custom labels and tooltip}
\usage{
mod_treemap_onelevel_custom_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap with custom labels and tooltip
}
