% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_onelevel_nocolors.R
\name{mod_treemap_onelevel_nocolors_plot}
\alias{mod_treemap_onelevel_nocolors_plot}
\title{Treemap: Share of Pokemon by main type (squarify)}
\usage{
mod_treemap_onelevel_nocolors_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap: Share of Pokemon by main type (squarify)
}
