% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_style.R
\name{mod_treemap_style_plot}
\alias{mod_treemap_style_plot}
\title{Treemap: Share of Pokemon by main type (styled)}
\usage{
mod_treemap_style_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap: Share of Pokemon by main type (styled)
}
