% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar3.R
\name{mod_bar3_plot}
\alias{mod_bar3_plot}
\title{Bar chart: Stacked Vertical Bars (attack/defense proportions)}
\usage{
mod_bar3_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Bar chart: Stacked Vertical Bars (attack/defense proportions)
}
