% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar4.R
\name{mod_bar4_plot}
\alias{mod_bar4_plot}
\title{Bar4 plot module (stacked horizontal)}
\usage{
mod_bar4_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Bar4 plot module (stacked horizontal)
}
