% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box4.R
\name{mod_box4_plot}
\alias{mod_box4_plot}
\title{Box4 plot module (log height)}
\usage{
mod_box4_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Box4 plot module (log height)
}
