% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geomap1.R
\name{mod_geomap1_plot}
\alias{mod_geomap1_plot}
\title{Geomap1 plot module}
\usage{
mod_geomap1_plot(data)
}
\arguments{
\item{data}{data.frame of region ids}
}
\value{
d3po widget
}
\description{
Geomap1 plot module
}
