% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geomap2.R
\name{mod_geomap2_plot}
\alias{mod_geomap2_plot}
\title{Geomap: Chile}
\usage{
mod_geomap2_plot(data = NULL)
}
\arguments{
\item{data}{data.frame (optional)}
}
\value{
d3po widget
}
\description{
Geomap: Chile
}
