% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geomap_custom.R
\name{mod_geomap_custom_plot}
\alias{mod_geomap_custom_plot}
\title{Geomap custom tooltip module}
\usage{
mod_geomap_custom_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Geomap custom tooltip module
}
