% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_custom.R
\name{mod_network_custom_plot}
\alias{mod_network_custom_plot}
\title{Network custom tooltip module}
\usage{
mod_network_custom_plot(data)
}
\arguments{
\item{data}{graph}
}
\value{
d3po widget
}
\description{
Network custom tooltip module
}
