% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_fr.R
\name{mod_network_fr_plot}
\alias{mod_network_fr_plot}
\title{Network FR layout plot module}
\usage{
mod_network_fr_plot(data)
}
\arguments{
\item{data}{graph or nodes/edges}
}
\value{
d3po widget
}
\description{
Network FR layout plot module
}
