% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pie_custom.R
\name{mod_pie_custom_plot}
\alias{mod_pie_custom_plot}
\title{Pie custom module}
\usage{
mod_pie_custom_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Pie custom module
}
