% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter1.R
\name{mod_scatter1_plot}
\alias{mod_scatter1_plot}
\title{Scatter1 plot module}
\usage{
mod_scatter1_plot(data)
}
\arguments{
\item{data}{data.frame of pokemon}
}
\value{
d3po widget
}
\description{
Scatter1 plot module
}
