% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter2.R
\name{mod_scatter2_plot}
\alias{mod_scatter2_plot}
\title{Scatter: Log(Height) vs Log(Weight)}
\usage{
mod_scatter2_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Scatter: Log(Height) vs Log(Weight)
}
