% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter3.R
\name{mod_scatter3_plot}
\alias{mod_scatter3_plot}
\title{Scatter: With size encoding (inverse distance from mean)}
\usage{
mod_scatter3_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Scatter: With size encoding (inverse distance from mean)
}
