% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter4.R
\name{mod_scatter4_plot}
\alias{mod_scatter4_plot}
\title{Scatter: Log scales with size encoding}
\usage{
mod_scatter4_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Scatter: Log scales with size encoding
}
