% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_custom.R
\name{mod_scatter_custom_plot}
\alias{mod_scatter_custom_plot}
\title{Scatter custom tooltip}
\usage{
mod_scatter_custom_plot(data = d3po::pokemon)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Scatter custom tooltip
}
