% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap_style.R
\name{mod_treemap_style_plot}
\alias{mod_treemap_style_plot}
\title{Treemap style example module}
\usage{
mod_treemap_style_plot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
d3po widget
}
\description{
Treemap style example module
}
